Split large update function into another function
